# apps/api/v2/swagger/documentation.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v2/organizations/{orgId}/schedules": {
      "get": {
        "operationId": "OrganizationsSchedulesController_getOrganizationSchedules",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "The number of items to return",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "The number of items to skip",
            "example": 0,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchedulesOutput_2024_06_11"
                }
              }
            }
          }
        },
        "tags": [
          "Organizations Schedules"
        ]
      }
    },
    "/v2/organizations/{orgId}/users/{userId}/schedules": {
      "post": {
        "operationId": "OrganizationsSchedulesController_createUserSchedule",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateScheduleOutput_2024_06_11"
                }
              }
            }
          }
        },
        "tags": [
          "Organizations Schedules"
        ]
      },
      "get": {
        "operationId": "OrganizationsSchedulesController_getUserSchedules",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchedulesOutput_2024_06_11"
                }
              }
            }
          }
        },
        "tags": [
          "Organizations Schedules"
        ]
      }
    },
    "/v2/organizations/{orgId}/users/{userId}/schedules/{scheduleId}": {
      "get": {
        "operationId": "OrganizationsSchedulesController_getUserSchedule",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "scheduleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScheduleOutput_2024_06_11"
                }
              }
            }
          }
        },
        "tags": [
          "Organizations Schedules"
        ]
      },
      "patch": {
        "operationId": "OrganizationsSchedulesController_updateUserSchedule",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "scheduleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateScheduleOutput_2024_06_11"
                }
              }
            }
          }
        },
        "tags": [
          "Organizations Schedules"
        ]
      },
      "delete": {
        "operationId": "OrganizationsSchedulesController_deleteUserSchedule",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "scheduleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteScheduleOutput_2024_06_11"
                }
              }
            }
          }
        },
        "tags": [
          "Organizations Schedules"
        ]
      }
    },
...
    "/v2/schedules": {
      "post": {
        "operationId": "SchedulesController_2024_04_15_createSchedule",
        "summary": "Create user availability schedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleInput_2024_04_15"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateScheduleOutput_2024_04_15"
                }
              }
            }
          }
        },
        "tags": [
          "Schedules"
        ]
      },
      "get": {
        "operationId": "SchedulesController_2024_04_15_getSchedules",
        "summary": "Get availability schedule",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchedulesOutput_2024_04_15"
                }
              }
            }
          }
        },
        "tags": [
          "Schedules"
        ]
      }
    },
...
    "/v2/schedules/{scheduleId}": {
      "get": {
        "operationId": "SchedulesController_2024_04_15_getSchedule",
        "summary": "Get availability schedule by ID",
        "parameters": [
          {
            "name": "scheduleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScheduleOutput_2024_04_15"
                }
              }
            }
          }
        },
        "tags": [
          "Schedules"
        ]
      },
      "patch": {
        "operationId": "SchedulesController_2024_04_15_updateSchedule",
        "summary": "Edit availability schedule by ID",
        "parameters": [
          {
            "name": "scheduleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateScheduleOutput_2024_04_15"
                }
              }
            }
          }
        },
        "tags": [
          "Schedules"
        ]
      },
      "delete": {
        "operationId": "SchedulesController_2024_04_15_deleteSchedule",
        "summary": "Delete availability schedule by ID",
        "parameters": [
          {
            "name": "scheduleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteScheduleOutput_2024_04_15"
                }
              }
            }
          }
        },
        "tags": [
          "Schedules"
        ]
      }
    },
...
  },
  "info": {
    "title": "OneHash Cal v2 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ScheduleAvailabilityInput_2024_06_11": {
        "type": "object",
        "properties": {
          "days": {
            "example": [
              "Monday",
              "Tuesday"
            ],
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "startTime": {
            "type": "string",
            "pattern": "TIME_FORMAT_HH_MM",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "pattern": "TIME_FORMAT_HH_MM",
            "example": "10:00"
          }
        },
        "required": [
          "days",
          "startTime",
          "endTime"
        ]
      },
      "ScheduleOverrideInput_2024_06_11": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2024-05-20"
          },
          "startTime": {
            "type": "string",
            "pattern": "TIME_FORMAT_HH_MM",
            "example": "12:00"
          },
          "endTime": {
            "type": "string",
            "pattern": "TIME_FORMAT_HH_MM",
            "example": "13:00"
          }
        },
        "required": [
          "date",
          "startTime",
          "endTime"
        ]
      },
      "ScheduleOutput_2024_06_11": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 254
          },
          "ownerId": {
            "type": "number",
            "example": 478
          },
          "name": {
            "type": "string",
            "example": "One-on-one coaching"
          },
          "timeZone": {
            "type": "string",
            "example": "Europe/Rome"
          },
          "availability": {
            "example": [
              {
                "days": [
                  "Monday",
                  "Tuesday"
                ],
                "startTime": "09:00",
                "endTime": "10:00"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleAvailabilityInput_2024_06_11"
            }
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          },
          "overrides": {
            "example": [
              {
                "date": "2024-05-20",
                "startTime": "12:00",
                "endTime": "13:00"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleOverrideInput_2024_06_11"
            }
          }
        },
        "required": [
          "id",
          "ownerId",
          "name",
          "timeZone",
          "availability",
          "isDefault",
          "overrides"
        ]
      },
      "GetSchedulesOutput_2024_06_11": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleOutput_2024_06_11"
            }
          },
          "error": {
            "type": "object"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "CreateScheduleOutput_2024_06_11": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleOutput_2024_06_11"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "GetScheduleOutput_2024_06_11": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ScheduleOutput_2024_06_11"
              }
            ]
          },
          "error": {
            "type": "object"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "UpdateScheduleOutput_2024_06_11": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleOutput_2024_06_11"
          },
          "error": {
            "type": "object"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "DeleteScheduleOutput_2024_06_11": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateAvailabilityInput_2024_04_15": {
        "type": "object",
        "properties": {
          "days": {
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "days",
          "startTime",
          "endTime"
        ]
      },
      "CreateScheduleInput_2024_04_15": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "availabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateAvailabilityInput_2024_04_15"
            }
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "timeZone",
          "isDefault"
        ]
      },
      "WorkingHours": {
        "type": "object",
        "properties": {
          "days": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "startTime": {
            "type": "number"
          },
          "endTime": {
            "type": "number"
          },
          "userId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "days",
          "startTime",
          "endTime"
        ]
      },
      "AvailabilityModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "number",
            "nullable": true
          },
          "eventTypeId": {
            "type": "number",
            "nullable": true
          },
          "days": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "scheduleId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "days",
          "startTime",
          "endTime"
        ]
      },
      "TimeRange": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "nullable": true
          },
          "start": {
            "format": "date-time",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ]
      },
      "ScheduleOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "isManaged": {
            "type": "boolean"
          },
          "workingHours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingHours"
            }
          },
          "schedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailabilityModel"
            }
          },
          "availability": {
            "type": "array",
            "items": {
              "required": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeRange"
              }
            }
          },
          "timeZone": {
            "type": "string"
          },
          "dateOverrides": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "isDefault": {
            "type": "boolean"
          },
          "isLastSchedule": {
            "type": "boolean"
          },
          "readOnly": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "isManaged",
          "workingHours",
          "schedule",
          "availability",
          "timeZone",
          "dateOverrides",
          "isDefault",
          "isLastSchedule",
          "readOnly"
        ]
      },
      "CreateScheduleOutput_2024_04_15": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleOutput"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
      "GetScheduleOutput_2024_04_15": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleOutput"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "GetSchedulesOutput_2024_04_15": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleOutput"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "EventTypeModel_2024_04_15": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "eventName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "AvailabilityModel_2024_04_15": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "number",
            "nullable": true
          },
          "scheduleId": {
            "type": "number",
            "nullable": true
          },
          "eventTypeId": {
            "type": "number",
            "nullable": true
          },
          "days": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "days"
        ]
      },
      "ScheduleModel_2024_04_15": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "timeZone": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventTypeModel_2024_04_15"
            }
          },
          "availability": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailabilityModel_2024_04_15"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "name"
        ]
      },
      "UpdatedScheduleOutput_2024_04_15": {
        "type": "object",
        "properties": {
          "schedule": {
            "$ref": "#/components/schemas/ScheduleModel_2024_04_15"
          },
          "isDefault": {
            "type": "boolean"
          },
          "timeZone": {
            "type": "string"
          },
          "prevDefaultId": {
            "type": "number",
            "nullable": true
          },
          "currentDefaultId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "schedule",
          "isDefault"
        ]
      },
      "UpdateScheduleOutput_2024_04_15": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/UpdatedScheduleOutput_2024_04_15"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "DeleteScheduleOutput_2024_04_15": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}